Define IPC objects and InvokeObject() functionality
Added support for InvokeObject() API to allow invocation of objects in
remote QTEE domain.

A remote object in QTEE represents a service which can be requested by the
client by invoking the object. As a result of the invocation, the client
can recieve another remote object to invoke.

Such an object-exchange between the client and QTEE can be initiated by
first obtaining a root object via the GetRootObject() API.

Signed-off-by: Harshal Dev <quic_hdev@quicinc.com>