Add support for Callback Objects and Supplicant
This change adds support for multi-threaded Callback Supplicant and
Callback Objects.

Clients can now define their own Callback Objects to host services for
QTEE, and call RegisterCallbackObject() to register them. When the client
sends a callback object to QTEE, QTEE can invoke it to generate a callback
request for the service.

Callback requests from QTEE are served by the Callback Supplicant, which
is part of the Root Object. Hence, it is initialized along with the Root
Object and shares the same TEE context.

Signed-off-by: Harshal Dev <quic_hdev@quicinc.com>